Added some old code back in, which allows the login window to kinda sorta function.
Need to clean up the old code. Most of it is commented out now.

Modified the "generateToken" function to "verifyCredentials", and now calling that from the loginWindow.

Need to fix the completion handler, so the UI functions properly, and also parse out the token/expiry/credentials from a successful run (which it is detecting).